cloudfunctions/verify-firebase-token/requirements.txt
# Function dependencies, for example:
# package>=version
google-auth==1.31.0
requests==2.21.0